+++ packages/winnow/CHANGELOG.md
@@ -2,8 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+* OutSR parameter handles latestWkid and wkt
+
 ## [1.5.0] - 02-10-2017
-###
+### Added
 * Support for geometry projections and the esri `outSR` parameter
 
 ## [1.4.2] - 02-06-2017